modified for feedback

+++ Src/Mcl/Dio_Types.c
@@ -9,28 +9,28 @@ Dio_LevelType Dio_ReadChannal (Dio_ChannelType ChannalId)
     uint8 pinnum = ChannalId%8  ;
     if (Portnum==0)
     {
-        return PORTA_REGS->GPIODATA[0x1<<pinnum] ;
+        return (uint8) PORTA_REGS->GPIODATA[0x1<<pinnum] ;
     }
     else if (Portnum==1)
     {
-        return PORTB_REGS->GPIODATA[0x1<<pinnum] ;
+        return (uint8)PORTB_REGS->GPIODATA[0x1<<pinnum] ;
 
     }
     else if (Portnum==2)
     {
-        return PORTC_REGS->GPIODATA[0x1<<pinnum] ;
+        return (uint8)PORTC_REGS->GPIODATA[0x1<<pinnum] ;
     }
     else if (Portnum==3)
     {
-        return PORTD_REGS->GPIODATA[0x1<<pinnum] ;
+        return (uint8)PORTD_REGS->GPIODATA[0x1<<pinnum] ;
     }
     else if (Portnum==4)
     {
-        return PORTE_REGS->GPIODATA[0x1<<pinnum] ;
+        return (uint8)PORTE_REGS->GPIODATA[0x1<<pinnum] ;
     }
     else
     {
-        return PORTF_REGS->GPIODATA[0x1<<pinnum] ;
+        return (uint8)PORTF_REGS->GPIODATA[0x1<<pinnum] ;
     }  
 }
 
@@ -42,28 +42,28 @@ void Dio_WriteChannal (Dio_ChannelType ChannalId,Dio_LevelType Level)
     uint8 pinnum = ChannalId%8  ;
     if (Portnum==0)
     {
-        PORTA_REGS->GPIODATA[0x1<<pinnum]=Level<<pinnum ;
+        PORTA_REGS->GPIODATA[0x1<<pinnum]=(uint32)(Level<<pinnum) ;
     }
     else if (Portnum==1)
     {
-        PORTB_REGS->GPIODATA[0x1<<pinnum]=Level<<pinnum ;
+        PORTB_REGS->GPIODATA[0x1<<pinnum]=(uint32)Level<<pinnum ;
 
     }
     else if (Portnum==2)
     {
-        PORTC_REGS->GPIODATA[0x1<<pinnum]=Level<<pinnum ;
+        PORTC_REGS->GPIODATA[0x1<<pinnum]=(uint32)Level<<pinnum ;
     }
     else if (Portnum==3)
     {
-        PORTD_REGS->GPIODATA[0x1<<pinnum]=Level<<pinnum ;
+        PORTD_REGS->GPIODATA[0x1<<pinnum]=(uint32)Level<<pinnum ;
     }
     else if (Portnum==4)
     {
-        PORTE_REGS->GPIODATA[0x1<<pinnum]=Level<<pinnum ;
+        PORTE_REGS->GPIODATA[0x1<<pinnum]=(uint32)Level<<pinnum ;
     }
     else
     {
-        PORTF_REGS->GPIODATA[0x1<<pinnum]=Level<<pinnum ;
+        PORTF_REGS->GPIODATA[0x1<<pinnum]=(uint32)Level<<pinnum ;
     }  
 
 }
@@ -73,28 +73,28 @@ Dio_PortLevelType Dio_ReadPort (Dio_PortType PortId)
 {
     if (PortId==0)
     {
-        return PORTA_REGS->GPIODATA[255] ;
+        return (uint8)PORTA_REGS->GPIODATA[255] ;
     }
     else if (PortId==1)
     {
-        return PORTB_REGS->GPIODATA[255] ;
+        return (uint8)PORTB_REGS->GPIODATA[255] ;
 
     }
     else if (PortId==2)
     {
-        return PORTC_REGS->GPIODATA[255] ;
+        return (uint8)PORTC_REGS->GPIODATA[255] ;
     }
     else if (PortId==3)
     {
-        return PORTD_REGS->GPIODATA[255] ;
+        return (uint8)PORTD_REGS->GPIODATA[255] ;
     }
     else if (PortId==4)
     {
-        return PORTE_REGS->GPIODATA[255] ;
+        return (uint8)PORTE_REGS->GPIODATA[255] ;
     }
     else
     {
-        return PORTF_REGS->GPIODATA[255] ;
+        return (uint8)PORTF_REGS->GPIODATA[255] ;
     }     
 }
 
@@ -134,38 +134,38 @@ void Dio_WritePort (Dio_PortType PortId ,Dio_PortLevelType Level)
 Dio_LevelType Dio_FlipChannal (Dio_ChannelType ChannalId)
 {
     Dio_PortType Portnum = ChannalId/8 ;
-    uint8 pinnum = ChannalId%8  ;
-		uint8 mask=1<<pinnum ;
+    uint32 pinnum = ChannalId%8  ;
+		uint32 mask=1<<pinnum ;
     if (Portnum==0)
     {
 			PORTA_REGS->GPIODATA[mask] ^= 0xFF;
-			return PORTA_REGS->GPIODATA[mask] ;
+			return (uint8)PORTA_REGS->GPIODATA[mask] ;
         
     }
     else if (Portnum==1)
     {
  			PORTB_REGS->GPIODATA[mask] ^= 0xFF;
-			return PORTB_REGS->GPIODATA[mask] ;
+			return (uint8)PORTB_REGS->GPIODATA[mask] ;
 
     }
     else if (Portnum==2)
     {
   		PORTC_REGS->GPIODATA[mask] ^= 0xFF;
-			return PORTC_REGS->GPIODATA[mask] ;
+			return (uint8)PORTC_REGS->GPIODATA[mask] ;
     }
     else if (Portnum==3)
     {
  			PORTD_REGS->GPIODATA[mask] ^= 0xFF;
-			return PORTD_REGS->GPIODATA[mask] ;
+			return (uint8)PORTD_REGS->GPIODATA[mask] ;
     }
     else if (Portnum==4)
     {
 			PORTE_REGS->GPIODATA[mask] ^= 0xFF;
-			return PORTE_REGS->GPIODATA[mask] ;
+			return (uint8)PORTE_REGS->GPIODATA[mask] ;
     }
     else
     {
  			PORTF_REGS->GPIODATA[mask] ^= 0xFF;
-			return PORTF_REGS->GPIODATA[mask] ;
+			return (uint8)PORTF_REGS->GPIODATA[mask] ;
     }
 }
